refactor and fix http modules
