+++ src/app/api/data-mgt/http/entity-impl/products.http-data-api.service.ts
@@ -9,11 +9,11 @@ import { Observable } from 'rxjs';
 import { Product } from 'src/app/models/entities/Product';
 import { ProductFilters } from 'src/app/shared/product-filters-panel/product-filters-panel.component';
 import { EntityDataApiIService } from '../../entity-data-api.iservice';
-import { EntityHttpDataApiService } from '../entity.http-data-api.aservice';
+import { HttpDataApiService } from '../http-data-api.aservice';
 
 @Injectable()
 export class ProductsHttpDataApiService
-  extends EntityHttpDataApiService
+  extends HttpDataApiService
   implements EntityDataApiIService<Product> {
 
   constructor(
