+++ src/app/api/data-mgt/http/entity-impl/sales.http-data-api.service.ts
@@ -10,11 +10,11 @@ import { map } from 'rxjs/operators';
 import { Sell } from 'src/app/models/entities/Sell';
 import { SellDetail } from 'src/app/models/entities/SellDetail';
 import { CompositeEntityDataApiIService } from '../../composite-entity-data-api.iservice';
-import { EntityHttpDataApiService } from '../entity.http-data-api.aservice';
+import { HttpDataApiService } from '../http-data-api.aservice';
 
 @Injectable()
 export class SalesHttpDataApiService
-  extends EntityHttpDataApiService
+  extends HttpDataApiService
   implements CompositeEntityDataApiIService<Sell, SellDetail> {
 
   constructor(
