+++ src/app/api/data-mgt/http/entity-impl/users.http-data-api.service.ts
@@ -7,12 +7,12 @@ import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
 import { User } from 'src/app/models/entities/User';
-import { EntityHttpDataApiService } from '../entity.http-data-api.aservice';
+import { HttpDataApiService } from '../http-data-api.aservice';
 import { EntityDataApiIService } from '../../entity-data-api.iservice';
 
 @Injectable()
 export class UsersHttpDataApiService
-  extends EntityHttpDataApiService
+  extends HttpDataApiService
   implements EntityDataApiIService<User> {
 
   constructor(
